src/bridge/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
qdrant-client>=1.12
openai>=1.50
litellm>=1.50
pydantic>=2.9
python-dotenv>=1.0
httpx>=0.27
pypdf>=5.0
python-docx>=1.1
beautifulsoup4>=4.12
langfuse>=2.50
sentence-transformers>=3.3
python-json-logger>=3.1
prometheus-fastapi-instrumentator>=7.0
